picar/requirements.txt
Cython==0.29.6
numpy==1.16.2
imutils==0.5.3
SpeechRecognition==3.8.1
opencv_contrib_python==4.1.1.26
keyboard==0.13.4
Adafruit_PCA9685==1.0.1
cryptography==2.8
ipaddr==2.2.0
lxml==4.4.2
ordereddict==1.1
picamera==1.13
protobuf==3.11.1
pyOpenSSL==19.1.0
pyzmq==18.1.1
wincertstore==0.2
